requirements.txt
# =====================================================
# Goat Farm Pro — Python Dependencies
# =====================================================
# Install karne ke liye:
#   pip install -r requirements.txt

# Core Framework
Django>=4.2,<5.2
django-ninja>=1.3.0

# CORS Support
django-cors-headers>=4.3.0

# Environment Variables (.env file support)
python-dotenv>=1.0.0

# HTTP Requests (Weather API ke liye)
requests>=2.31.0

# Excel Import / Export
openpyxl>=3.1.0
pandas>=2.0.0
xlrd>=2.0.1

# Image Handling (goat photos etc.)
Pillow>=10.1.0

# Date / Time Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# ==================== v6.0 New Dependencies ====================
# Batch 1
reportlab>=4.0.0          # Invoice PDF generation
# Batch 2
qrcode[pil]>=7.4.2        # QR Code generation per goat
twilio>=8.0.0             # WhatsApp/SMS notifications
# Batch 2 - AI Engine
numpy>=1.24.0             # Numerical calculations
# Batch 3 (future)
# celery>=5.3.0           # Background tasks (for cron reminders)
# redis>=5.0.0            # Celery broker
# django-guardian>=2.4.0  # Object-level permissions
